backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
PyJWT
pymongo
python-dotenv
PyPDF2
python-docx
requests
Werkzeug